requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.5.5  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower
uvicorn[standard]==0.22.0  # https://github.com/encode/uvicorn
hupper

# Django
# ------------------------------------------------------------------------------
django==4.1.8  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-redis==5.2.0  # https://github.com/jazzband/django-redis

# API-Related
# ------------------------------------------------------------------------------
django-ninja==0.21.0  # https://github.com/vitalik/django-ninja
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
djangorestframework-api-key==2.*  # https://github.com/florimondmanca/djangorestframework-api-key
django-ninja-jwt==5.2.5  # https://github.com/eadwinCode/django-ninja-jwt
django-ninja-extra
django-cors-headers==3.14.0
# Websockets
# ------------------------------------------------------------------------------
channels[daphne]
channels_redis

# NLP-Related
# ------------------------------------------------------------------------------
# llama_index==0.5.25  # https://github.com/jerryjliu/llama_index
llama_index==0.7.9  # https://github.com/jerryjliu/llama_index
PyPDF2==3.*  # https://pypdf2.readthedocs.io/en/latest/
pypdf==3.12.0
docx2txt==0.8
nebula3-python==3.4.0
